chore: fixed linting
Fixed linting and also implemented a better handling for safe functions
to always print to stderr and then exit.

Added strings,join to make it so an array and a char can be given and it
will join the string.

UUID v4 is looking good. It is quite fast and generates valid UUID.

Signed-off-by: Eric Bode <eric.bode@foundries.io>